Create POST /auth/login api endpoint
API to login a user will be created. A new token
will be generated on login, the Authorization filed
will be set with the token in the header if login
is sucessful or else error codes and messages will be
sent. Rename model to class naming standards.
Use babel-node with every script and updated in package.json.
An utility function checkPassword will be created to
check the password from response and return boolean
depending on the outcome

-	modified:   .gitignore
-	new file:   server/src/controller2/loginUser.js
-	new file:   server/src/controller2/utils/checkPassword.js
-	deleted:    server/src/db/test.js