Create getting and creating requests functionality
feat(user page): the attached accordion will be removed and
and made into a reusable function. The api endpoint to create
a request and get all requests will be restructured and the
modal will be removed, form cleaned upon creation of requests
which will be added to the list of requests.

-	modified:   UI/index.html
-	deleted:    UI/scripts/accordion.js
-	modified:   UI/scripts/makeRequest.js
-	deleted:    UI/scripts/request-tab.js
-	modified:   UI/userpage.html
-	modified:   server/controller/utils/generateToken.js
-	modified:   server/middleware/authenticate.js

Delivers[158029801]