Refactor code and create edit request template
feat(user page): Fix- the script to make request is refactored
to clean up dead code. The data returned from controller to create a request
is restructured to be uniform with data returned for all user requests.
feat- The user requests frontend will have an edit button that redirects to
the editrequest page which is a form with the request to be edited.

-	new file:   UI/editrequest.html
-	new file:   UI/scripts/editRequest.js
-	modified:   UI/scripts/makeRequest.js
-	modified:   UI/userpage.html
-	modified:   server/controller/createUserRequest.js