Create frontend modify request functionality
feat(modify request): the controllers and middleware for creating
and decoding a user is edited to make the data uniform. New html
page is modified and form to update request is added and script
to update request that sends data to the server and redirects to the user page is added.

-	modified:   UI/editrequest.html
-	modified:   UI/scripts/editRequest.js
-	modified:   server/controller/getAllUserRequest.js
-	modified:   server/controller/updateUserRequest.js
-	modified:   server/controller/utils/generateToken.js
-	modified:   server/middleware/authenticate.js