Create frontend get all requests, approve and reject request
feat(admin page): the frontend functionality for admin to get all the application
requests will be handled by the displayAllRequests function in the script. The
function will be responsible for getting the requests from server, storing the
requests in localstorage, and dynamically populating the table body with the requests
returned.
approve request functionality will be handled by the approveRequest function and
reject request will be handled by the disApproveRequest function they will both be
executed when the accept or reject button is clicked respectively and this will
update the status of the request and disbale the button depending on the function.
The controller for both is edited to be able to check adminRole rather than previous
admin_role and likewise, the styles and html structure will be adjusted to easily get
tablebody child nodes and give a not-allowed cursor to the button when clicked.

-	modified:   UI/admin.html
-	modified:   UI/css/stylesheet.css
-	deleted:    UI/scripts/admin-request.js
-	new file:   UI/scripts/adminRequest.js
-	new file:   UI/scripts/displayAdminRequests.js
-	modified:   server/controller/approveRequest.js
-	modified:   server/controller/getAllRequest.js
-	modified:   server/controller/rejectRequest.js