Create UI function to filter requests
feat(admin page): the function to filter requests is added.
it uses an input event on the select tag to check when the
value has changed and uses that to change the URL to get the
requests and the displayAllRequests will be called afterwards.
the major idea of this function is to change URL to query URL
or none query URL when it notices the select value has changed.

-	modified:   UI/admin.html
-	modified:   UI/scripts/displayAdminRequests.js