Create tests to get all user requests and a request
chore(get requests endpoint): test to get all user requests
endpoint is created which checks if the user is valid, has a
token or not and if the user has requests created or not.
Get a user request is created and checks if the id of request
is created by user and when not created by the user.

-	modified:   server/controller/getUserRequest.js
-	modified:   server/tests/api.test.js
-	modified:   server/tests/seed/seed.js

Delivers [157561804, 157620645]