Create tests for modify request api endpoint
chore(update request api): tests to check when the request
field is empty, missing a field, if the request did not
create the particular request, to not allow a request not pending
to be modified and also modify request will be created.
updateUserRequest controller will be modified to check for
user_id also when querying the database.

-	modified:   server/controller/updateUserRequest.js
-	modified:   server/tests/api.test.js
-	modified:   server/tests/seed/seed.js

Delivers [157627001]