Add and fix frontend UI and UX
fix(UI): the header and navbar all the pages will be combined and logo will be moved
to left. signup and login pages text will be removed and the form will be aligned to
center of the viewport.
The make a request function is placed in the header on userpage and the background
colour of the requests tabs when active will be made changed, the footer will be
removed aslo on admin and editrequest pages.

-	modified:   UI/admin.html
-	modified:   UI/css/stylesheet.css
-	modified:   UI/editrequest.html
-	modified:   UI/index.html
-	modified:   UI/login.html
-	modified:   UI/signup.html
-	modified:   UI/userpage.html