Add validation and display username
feat(user page): validation using submit event is added which will not allow an empty input field before submition. The header is made to display the username on signup and
login, the username name is saved in localstorage to be accessed by other pages.
A text is added which will be hidden when the request list is not empty and
displayed when the request is empty for the user.

-	modified:   UI/css/stylesheet.css
-	modified:   UI/scripts/login.js
-	modified:   UI/scripts/makeRequest.js
-	modified:   UI/scripts/signup.js
-	modified:   UI/userpage.html