Add error popup on request creation form
feat(create request): a popup text is addded to the form which displays when there is
error in the request. A function is added to display the popup and change event is used
to listen when the form is changed in response to error popup.
The signup and login api endpoint controller is edited to reposition the point of
trimming inputs in the program execution.

-	modified:   UI/css/stylesheet.css
-	modified:   UI/scripts/makeRequest.js
-	modified:   UI/userpage.html
-	modified:   server/controller/createAccount.js
-	modified:   server/controller/createUserRequest.js
-	modified:   server/controller/loginUser.js