Add error message to be displayed
feat(edit request): paragraph tag to display error message on updating request is
added, function to show error message is added. the style of the error message is
added, the click event on update button is changed to submit event to allow form
validation. Friendly error message is added if fetching API fails to be displayed to
user.

-	modified:   UI/css/stylesheet.css
-	modified:   UI/editrequest.html
-	modified:   UI/scripts/editRequest.js